requirements.txt
langchain
openai
streamlit
python-dotenv
wikipedia
youtube-transcript-api
faiss-cpu
docx2txt
gdown
pypdf
